add document_types module and move functionality there
