+++ corehq/apps/change_feed/topics.py
@@ -1,9 +1,9 @@
-from couchforms.models import all_known_formlike_doc_types
+from .document_types import CASE, FORM, META
 
-
-CASE = 'case'
-FORM = 'form'
-META = 'meta'
+# this is redundant but helps avoid import warnings until nothing references these
+CASE = CASE
+FORM = FORM
+META = META
 
 # new models
 CASE_SQL = 'case-sql'
@@ -16,11 +16,5 @@ ALL = (
 )
 
 
-def get_topic(document_type):
-    if document_type in ('CommCareCase', 'CommCareCase-Deleted'):
-        return CASE
-    elif document_type in all_known_formlike_doc_types():
-        return FORM
-    else:
-        # at some point we may want to make this more granular
-        return META
+def get_topic(document_type_object):
+    return document_type_object.primary_type
